unsigned int is too small for storing huge corpora (and one PatternPointer can be one corpus), using size_t instead #41

+++ include/pattern.h
@@ -97,8 +97,8 @@ class Pattern {
       * @param begin Index of the first token to copy (0-indexed)
       * @param length Number of tokens to copy
       */
-     Pattern(const Pattern& ref, unsigned int begin, unsigned int length, unsigned int * byteoffset=NULL, bool byteoffset_shiftone = false);
-     Pattern(const PatternPointer& ref,unsigned int begin, unsigned int length, unsigned int * byteoffset=NULL, bool byteoffset_shiftone = false);
+     Pattern(const Pattern& ref, size_t begin, size_t length, size_t * byteoffset=NULL, bool byteoffset_shiftone = false);
+     Pattern(const PatternPointer& ref,size_t begin, size_t length, size_t * byteoffset=NULL, bool byteoffset_shiftone = false);
 
      /**
       * Copy constructor for Pattern
@@ -427,9 +427,9 @@ class PatternPointer {
      void write(std::ostream * out, const unsigned char * corpusstart = NULL) const;
 
      //slice construtors:
-     PatternPointer(unsigned char *, unsigned int,unsigned int, unsigned int * byteoffset=NULL, bool byteoffset_shiftone = false);
-     PatternPointer(const PatternPointer&, unsigned int,unsigned int, unsigned int * byteoffset=NULL, bool byteoffset_shiftone = false);
-     PatternPointer(const Pattern&, unsigned int,unsigned int, unsigned int * byteoffset=NULL, bool byteoffset_shiftone = false);
+     PatternPointer(unsigned char *, size_t,size_t, size_t * byteoffset=NULL, bool byteoffset_shiftone = false);
+     PatternPointer(const PatternPointer&, size_t,size_t, size_t * byteoffset=NULL, bool byteoffset_shiftone = false);
+     PatternPointer(const Pattern&, size_t,size_t, size_t * byteoffset=NULL, bool byteoffset_shiftone = false);
 
      uint32_t computemask() const;
      uint32_t getmask() const { return mask; }
